+++ vtm/src/org/oscim/renderer/bucket/LineBucket.java
@@ -28,7 +28,6 @@ import org.oscim.renderer.GLState;
 import org.oscim.renderer.GLUtils;
 import org.oscim.renderer.GLViewport;
 import org.oscim.theme.styles.LineStyle;
-import org.oscim.utils.Parameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -665,7 +664,7 @@ public class LineBucket extends RenderBucket {
                     blur = false;
                 }
 
-                if (Parameters.TRANSPARENT_LINES && !Color.isOpaque(line.color)) {
+                if (line.transparent && !Color.isOpaque(line.color)) {
                     gl.depthMask(true);
                     gl.clear(GL.DEPTH_BUFFER_BIT);
                     GLState.test(true, false);
@@ -712,7 +711,7 @@ public class LineBucket extends RenderBucket {
                     gl.drawArrays(GL.TRIANGLE_STRIP,
                             b.vertexOffset, b.numVertices);
 
-                    if (Parameters.TRANSPARENT_LINES && !Color.isOpaque(line.color)) {
+                    if (line.transparent && !Color.isOpaque(line.color)) {
                         gl.depthMask(false);
                     }
 
@@ -763,7 +762,7 @@ public class LineBucket extends RenderBucket {
                             ref.vertexOffset, ref.numVertices);
                 }
 
-                if (Parameters.TRANSPARENT_LINES && !Color.isOpaque(line.color)) {
+                if (line.transparent && !Color.isOpaque(line.color)) {
                     gl.depthMask(false);
                 }
             }